refactor: Move providers to lib folder
* Move `samcli/commands/local/lib/*providers.py` to `samcli/lib/providers/`
* Update imports
* Update patch annotations
* Move `InvalidLayerReference` to an exception.py file within the destination
