+++ samcli/commands/local/invoke/cli.py
@@ -110,7 +110,7 @@ def do_cli(  # pylint: disable=R0914
     """
 
     from samcli.commands.exceptions import UserException
-    from samcli.commands.local.lib.exceptions import InvalidLayerReference
+    from samcli.lib.providers.exceptions import InvalidLayerReference
     from samcli.commands.local.cli_common.invoke_context import InvokeContext
     from samcli.local.lambdafn.exceptions import FunctionNotFound
     from samcli.commands.validate.lib.exceptions import InvalidSamDocumentException
