+++ samcli/commands/local/lib/exceptions.py
@@ -13,14 +13,3 @@ class OverridesNotWellDefinedError(Exception):
     """
     Raised when the overrides file is invalid
     """
-
-
-class InvalidLayerReference(Exception):
-    """
-    Raised when the LayerVersion LogicalId does not exist in the template
-    """
-
-    def __init__(self):
-        super(InvalidLayerReference, self).__init__(
-            "Layer References need to be of type " "'AWS::Serverless::LayerVersion' or 'AWS::Lambda::LayerVersion'"
-        )
