+++ samcli/commands/local/start_api/cli.py
@@ -110,7 +110,8 @@ def do_cli(  # pylint: disable=R0914
     """
 
     from samcli.commands.local.cli_common.invoke_context import InvokeContext
-    from samcli.commands.local.lib.exceptions import NoApisDefined, InvalidLayerReference
+    from samcli.commands.local.lib.exceptions import NoApisDefined
+    from samcli.lib.providers.exceptions import InvalidLayerReference
     from samcli.commands.exceptions import UserException
     from samcli.commands.local.lib.local_api_service import LocalApiService
     from samcli.commands.validate.lib.exceptions import InvalidSamDocumentException
