+++ samcli/local/apigw/local_apigw_service.py
@@ -7,7 +7,7 @@ import base64
 from flask import Flask, request
 from werkzeug.datastructures import Headers
 
-from samcli.commands.local.lib.provider import Cors
+from samcli.lib.providers.provider import Cors
 from samcli.local.services.base_local_service import BaseLocalService, LambdaOutputParser
 from samcli.lib.utils.stream_writer import StreamWriter
 from samcli.local.lambdafn.exceptions import FunctionNotFound
