+++ tests/unit/commands/local/invoke/test_cli.py
@@ -7,7 +7,7 @@ from unittest.mock import patch, Mock
 from parameterized import parameterized, param
 
 from samcli.local.lambdafn.exceptions import FunctionNotFound
-from samcli.commands.local.lib.exceptions import InvalidLayerReference
+from samcli.lib.providers.exceptions import InvalidLayerReference
 from samcli.commands.validate.lib.exceptions import InvalidSamDocumentException
 from samcli.commands.exceptions import UserException
 from samcli.commands.local.invoke.cli import do_cli as invoke_cli, _get_event as invoke_cli_get_event
