+++ tests/unit/commands/local/lib/test_sam_base_provider.py
@@ -1,12 +1,12 @@
 from unittest import TestCase
 from unittest.mock import Mock, patch
-from samcli.commands.local.lib.sam_base_provider import SamBaseProvider
+from samcli.lib.providers.sam_base_provider import SamBaseProvider
 from samcli.lib.intrinsic_resolver.intrinsic_property_resolver import IntrinsicResolver
 
 
 class TestSamBaseProvider_get_template(TestCase):
-    @patch("samcli.commands.local.lib.sam_base_provider.ResourceMetadataNormalizer")
-    @patch("samcli.commands.local.lib.sam_base_provider.SamTranslatorWrapper")
+    @patch("samcli.lib.providers.sam_base_provider.ResourceMetadataNormalizer")
+    @patch("samcli.lib.providers.sam_base_provider.SamTranslatorWrapper")
     @patch.object(IntrinsicResolver, "resolve_template")
     def test_must_run_translator_plugins(
         self, resolve_template_mock, SamTranslatorWrapperMock, resource_metadata_normalizer_patch
